+++ illuin_inject/bindings/registered_binding.py
@@ -7,11 +7,11 @@ from illuin_inject.typings import InjectedT
 from .binding import Binding
 
 if TYPE_CHECKING:
-    from .abstract_binding_spec import AbstractBindingSpec
+    from .abstract_module import AbstractModule
 
 
 @attr.s(auto_attribs=True, frozen=True)
 class RegisteredBinding(Generic[InjectedT]):
     raw_binding: Binding[InjectedT]
-    source_path: Tuple["AbstractBindingSpec", ...] = attr.Factory(tuple)
+    source_path: Tuple["AbstractModule", ...] = attr.Factory(tuple)
     target: Target[InjectedT] = attr.Factory(lambda self: self.raw_binding.target, takes_self=True)