+++ tests/test_bindings/test_module.py
@@ -1,0 +1,12 @@
+import unittest
+
+from illuin_inject import Module
+
+
+class TestModule(unittest.TestCase):
+    def setUp(self) -> None:
+        self.module = Module()
+
+    def test_configure_is_not_implemented(self):
+        with self.assertRaises(NotImplementedError):
+            self.module.configure()